[build] change client structure
