+++ client/src/components/tool/toolList.js
@@ -2,9 +2,9 @@ import  Axios  from 'axios';
 import React, { useEffect, useState } from 'react';
 import Tab from 'react-bootstrap/Tab';
 import Tabs from 'react-bootstrap/Tabs';
-import lbs from '../labels';
+import lbs from '../../labels';
 import ToolCard from './toolCard';
-import Loading from './loading';
+import Loading from '../loading';
 
 const dummy = {
     "Domain":[
