+++ client/src/components/tool/toolResultBanner.js
