+++ client/src/components/visualization/timeline/suspectTimeline.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SuspectTimeline = () => {
+  return (
+    <div>SuspectTimeline</div>
+  )
+}
+
+export default SuspectTimeline
