+++ client/src/components/visualization/timeline/timeline.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import DomainTimeline from './domainTimeline'
+
+const Timeline = (props) => {
+  const isDone = props.isDone
+  return (
+    <div>
+      {/* <DomainTimeline isDone={isDone}/> */}
+    </div>
+  )
+}
+
+export default Timeline
