+++ client/src/components/visualization/timeline/wholeTimeline.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const WholeTimeline = () => {
+  return (
+    <div>WholeTimeline</div>
+  )
+}
+
+export default WholeTimeline
